maven walkmod plugin
This patch adds the walkmod maven plugin. It is executed after
compiling classes because walkmod requires to resolve the application
classpath. If walkmod fixes any checkstyle issue (syntactic or
semantic), it will generate a walkmod.patch file.

Notice that formatting issues requires to use the Eclipse formatter and thus,
it is not focused on Checkstyle rules.